Fix PostgreSQL port and password in config; update price handling to use Decimal in trade_routes

+++ routes/trade_routes.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from flask import Blueprint, jsonify, request,session
 from service.trade_service import place_order
 
@@ -14,7 +15,8 @@ def place_trade_order():
         quantity = int(quantity) if quantity else 0
         order_type = data.get("order_type")
         price = data.get("price")  
-        price = float(price) if price else None
+        price = Decimal(price) if price else None
+
         transaction_type = data.get("transaction_type")
 
         result = place_order(
